backend/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.0
django-filter==23.3
Pillow==10.1.0
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
whitenoise==6.6.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.11.2
django-storages==1.14.2
drf-spectacular==0.26.4
python-decouple==3.8
sentry-sdk==1.32.0 